+++ ebs-20210730/core/Models/DescribeEnterpriseSnapshotPolicyResponseBody.cs
@@ -10,7 +10,7 @@ namespace AlibabaCloud.SDK.Ebs20210730.Models
 {
     public class DescribeEnterpriseSnapshotPolicyResponseBody : TeaModel {
         /// <summary>
-        /// The query token returned in this call.
+        /// A pagination token. It can be used in the next request to retrieve a new page of results. If NextToken is empty, no next page exists.
         /// </summary>
         [NameInMap("NextToken")]
         [Validation(Required=false)]
@@ -31,53 +31,49 @@ namespace AlibabaCloud.SDK.Ebs20210730.Models
         public int? PageSize { get; set; }
 
         /// <summary>
-        /// The list of policies.
+        /// The returned enterprise-level snapshot policies.
         /// </summary>
         [NameInMap("Policies")]
         [Validation(Required=false)]
         public List<DescribeEnterpriseSnapshotPolicyResponseBodyPolicies> Policies { get; set; }
         public class DescribeEnterpriseSnapshotPolicyResponseBodyPolicies : TeaModel {
             /// <summary>
-            /// Creation Time.
+            /// The time when the enterprise-level snapshot policy was created.
             /// </summary>
             [NameInMap("CreateTime")]
             [Validation(Required=false)]
             public string CreateTime { get; set; }
 
             /// <summary>
-            /// Snapshot replication destination information.
+            /// The replication rule of snapshots in the enterprise-level snapshot policy.
             /// </summary>
             [NameInMap("CrossRegionCopyInfo")]
             [Validation(Required=false)]
             public DescribeEnterpriseSnapshotPolicyResponseBodyPoliciesCrossRegionCopyInfo CrossRegionCopyInfo { get; set; }
             public class DescribeEnterpriseSnapshotPolicyResponseBodyPoliciesCrossRegionCopyInfo : TeaModel {
                 /// <summary>
-                /// Whether cross-region replication is enabled. The range of values:
-                /// 
-                /// - true
-                /// 
-                /// - false
+                /// Indicates whether the cross-region replication feature is enabled.
                 /// </summary>
                 [NameInMap("Enabled")]
                 [Validation(Required=false)]
                 public bool? Enabled { get; set; }
 
                 /// <summary>
-                /// Destination region information.
+                /// The destination regions that store snapshot copies.
                 /// </summary>
                 [NameInMap("Regions")]
                 [Validation(Required=false)]
                 public List<DescribeEnterpriseSnapshotPolicyResponseBodyPoliciesCrossRegionCopyInfoRegions> Regions { get; set; }
                 public class DescribeEnterpriseSnapshotPolicyResponseBodyPoliciesCrossRegionCopyInfoRegions : TeaModel {
                     /// <summary>
-                    /// The region ID of the destination. You can call the [DescribeRegions](~~25609~~) operation to query the most recent region list.
+                    /// The ID of the destination region.
                     /// </summary>
                     [NameInMap("RegionId")]
                     [Validation(Required=false)]
                     public string RegionId { get; set; }
 
                     /// <summary>
-                    /// Number of days to retain the destination snapshot. The range of values is greater than 1.
+                    /// The retention period of snapshot copies in the destination region. Unit: day.
                     /// </summary>
                     [NameInMap("RetainDays")]
                     [Validation(Required=false)]
@@ -88,74 +84,70 @@ namespace AlibabaCloud.SDK.Ebs20210730.Models
             }
 
             /// <summary>
-            /// The description of the policy.
+            /// The description of the enterprise-level snapshot policy.
             /// </summary>
             [NameInMap("Desc")]
             [Validation(Required=false)]
             public string Desc { get; set; }
 
+            /// <summary>
+            /// The list of disks.
+            /// </summary>
             [NameInMap("DiskIds")]
             [Validation(Required=false)]
             public List<string> DiskIds { get; set; }
 
             /// <summary>
-            /// Indicates whether it is an ECS snapshot policy，The range of values:
-            /// 
-            /// - true
-            /// - false
+            /// Indicates whether snapshots are managed.
             /// </summary>
             [NameInMap("ManagedForEcs")]
             [Validation(Required=false)]
             public bool? ManagedForEcs { get; set; }
 
             /// <summary>
-            /// The name of the policy.
+            /// The name of the enterprise-level snapshot policy.
             /// </summary>
             [NameInMap("Name")]
             [Validation(Required=false)]
             public string Name { get; set; }
 
             /// <summary>
-            /// The id of the policy.
+            /// The ID of the enterprise-level snapshot policy.
             /// </summary>
             [NameInMap("PolicyId")]
             [Validation(Required=false)]
             public string PolicyId { get; set; }
 
             /// <summary>
-            /// The resource group
+            /// the resource group
             /// </summary>
             [NameInMap("ResourceGroupId")]
             [Validation(Required=false)]
             public string ResourceGroupId { get; set; }
 
             /// <summary>
-            /// The snapshot retention rule.
+            /// The retention rule of the enterprise-level snapshot policy.
             /// </summary>
             [NameInMap("RetainRule")]
             [Validation(Required=false)]
             public DescribeEnterpriseSnapshotPolicyResponseBodyPoliciesRetainRule RetainRule { get; set; }
             public class DescribeEnterpriseSnapshotPolicyResponseBodyPoliciesRetainRule : TeaModel {
                 /// <summary>
-                /// Maximum number of retained snapshots.
+                /// The maximum number of snapshots that can be retained.
                 /// </summary>
                 [NameInMap("Number")]
                 [Validation(Required=false)]
                 public int? Number { get; set; }
 
                 /// <summary>
-                /// The time interval , valid value greater than 1.
+                /// The value of the retention period of snapshots.
                 /// </summary>
                 [NameInMap("TimeInterval")]
                 [Validation(Required=false)]
                 public int? TimeInterval { get; set; }
 
                 /// <summary>
-                /// The unit of time, valid values:
-                /// 
-                /// - DAYS
-                /// 
-                /// - WEEKS
+                /// The unit of the retention period of snapshots.
                 /// </summary>
                 [NameInMap("TimeUnit")]
                 [Validation(Required=false)]
@@ -164,16 +156,14 @@ namespace AlibabaCloud.SDK.Ebs20210730.Models
             }
 
             /// <summary>
-            /// The rule for scheduling.
+            /// The scheduling rule of the enterprise-level snapshot policy.
             /// </summary>
             [NameInMap("Schedule")]
             [Validation(Required=false)]
             public DescribeEnterpriseSnapshotPolicyResponseBodyPoliciesSchedule Schedule { get; set; }
             public class DescribeEnterpriseSnapshotPolicyResponseBodyPoliciesSchedule : TeaModel {
                 /// <summary>
-                /// The time when the policy will to be scheduled. Valid values: Set the parameter in a cron expression.
-                /// 
-                /// For example, you can use 0 0 4 1/1 * ? to specify 04:00:00 (UTC+8) on the first day of each month.
+                /// The cron expression of the enterprise-level snapshot policy.
                 /// </summary>
                 [NameInMap("CronExpression")]
                 [Validation(Required=false)]
@@ -182,17 +172,14 @@ namespace AlibabaCloud.SDK.Ebs20210730.Models
             }
 
             /// <summary>
-            /// The special snapshot retention rules.
+            /// The special retention rules of the enterprise-level snapshot policy.
             /// </summary>
             [NameInMap("SpecialRetainRules")]
             [Validation(Required=false)]
             public DescribeEnterpriseSnapshotPolicyResponseBodyPoliciesSpecialRetainRules SpecialRetainRules { get; set; }
             public class DescribeEnterpriseSnapshotPolicyResponseBodyPoliciesSpecialRetainRules : TeaModel {
                 /// <summary>
-                /// Indicates whether the special retention is enabled.
-                /// 
-                /// *   true: enable
-                /// *   false: disable
+                /// Indicates whether the special retention period is enabled.
                 /// </summary>
                 [NameInMap("Enabled")]
                 [Validation(Required=false)]
@@ -206,28 +193,21 @@ namespace AlibabaCloud.SDK.Ebs20210730.Models
                 public List<DescribeEnterpriseSnapshotPolicyResponseBodyPoliciesSpecialRetainRulesRules> Rules { get; set; }
                 public class DescribeEnterpriseSnapshotPolicyResponseBodyPoliciesSpecialRetainRulesRules : TeaModel {
                     /// <summary>
-                    /// The periodic unit for specially retained snapshots. If configured to WEEKS, it provides special retention for the first snapshot of each week. The retention period is determined by TimeUnit and TimeInterval. The range of values are:
-                    /// - WEEKS
-                    /// - MONTHS
-                    /// - YEARS
+                    /// The unit of the special retention period.
                     /// </summary>
                     [NameInMap("SpecialPeriodUnit")]
                     [Validation(Required=false)]
                     public string SpecialPeriodUnit { get; set; }
 
                     /// <summary>
-                    /// Retention Time Value. The range of values is greater than 1.
+                    /// The value of the retention period.
                     /// </summary>
                     [NameInMap("TimeInterval")]
                     [Validation(Required=false)]
                     public int? TimeInterval { get; set; }
 
                     /// <summary>
-                    /// Retention time unit for special snapshots. The range of values:
-                    /// 
-                    /// - DAYS
-                    /// 
-                    /// - WEEKS
+                    /// The unit of the retention period.
                     /// </summary>
                     [NameInMap("TimeUnit")]
                     [Validation(Required=false)]
@@ -238,29 +218,21 @@ namespace AlibabaCloud.SDK.Ebs20210730.Models
             }
 
             /// <summary>
-            /// The status of the policy. Valid values:
-            /// 
-            /// - ENABLED: Enable policy execution.
-            /// 
-            /// - DISABLED: Disable policy execution.
+            /// The status of the enterprise-level snapshot policy.
             /// </summary>
             [NameInMap("State")]
             [Validation(Required=false)]
             public string State { get; set; }
 
             /// <summary>
-            /// Advanced snapshot features.
+            /// The storage rule of snapshots in the enterprise-level snapshot policy.
             /// </summary>
             [NameInMap("StorageRule")]
             [Validation(Required=false)]
             public DescribeEnterpriseSnapshotPolicyResponseBodyPoliciesStorageRule StorageRule { get; set; }
             public class DescribeEnterpriseSnapshotPolicyResponseBodyPoliciesStorageRule : TeaModel {
                 /// <summary>
-                /// Whether to enable the rapid availability of snapshots. The range of values:
-                /// 
-                /// - true
-                /// 
-                /// - false
+                /// Indicates whether the instant access feature is enabled.
                 /// </summary>
                 [NameInMap("EnableImmediateAccess")]
                 [Validation(Required=false)]
@@ -269,21 +241,21 @@ namespace AlibabaCloud.SDK.Ebs20210730.Models
             }
 
             /// <summary>
-            /// The list of tags.
+            /// the pair tags
             /// </summary>
             [NameInMap("Tags")]
             [Validation(Required=false)]
             public List<DescribeEnterpriseSnapshotPolicyResponseBodyPoliciesTags> Tags { get; set; }
             public class DescribeEnterpriseSnapshotPolicyResponseBodyPoliciesTags : TeaModel {
                 /// <summary>
-                /// The key of a tag.
+                /// The key of the tag of the enterprise-level snapshot policy.
                 /// </summary>
                 [NameInMap("TagKey")]
                 [Validation(Required=false)]
                 public string TagKey { get; set; }
 
                 /// <summary>
-                /// The value of a tag.
+                /// The value of the tag of the enterprise-level snapshot policy.
                 /// </summary>
                 [NameInMap("TagValue")]
                 [Validation(Required=false)]
@@ -292,14 +264,14 @@ namespace AlibabaCloud.SDK.Ebs20210730.Models
             }
 
             /// <summary>
-            /// Number of bound targets.
+            /// The number of objects that are associated with the enterprise-level snapshot policy.
             /// </summary>
             [NameInMap("TargetCount")]
             [Validation(Required=false)]
             public int? TargetCount { get; set; }
 
             /// <summary>
-            /// The target type.
+            /// The type of the enterprise-level snapshot policy.
             /// </summary>
             [NameInMap("TargetType")]
             [Validation(Required=false)]
@@ -308,7 +280,7 @@ namespace AlibabaCloud.SDK.Ebs20210730.Models
         }
 
         /// <summary>
-        /// The ID of the request.
+        /// The request ID.
         /// </summary>
         [NameInMap("RequestId")]
         [Validation(Required=false)]