+++ ebs-20210730/core/Models/QueryDedicatedBlockStorageClusterInventoryDataRequest.cs
@@ -1,0 +1,58 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.Ebs20210730.Models
+{
+    public class QueryDedicatedBlockStorageClusterInventoryDataRequest : TeaModel {
+        /// <summary>
+        /// The client token that is used to ensure the idempotence of the request. You can use the client to generate the value, but you must ensure that it is unique among different requests.
+        /// 
+        /// The ClientToken value can contain only ASCII characters and cannot exceed 64 characters in length. For more information, see [How do I ensure idempotence ](~~25693~~).
+        /// </summary>
+        [NameInMap("ClientToken")]
+        [Validation(Required=false)]
+        public string ClientToken { get; set; }
+
+        /// <summary>
+        /// The ID of the dedicated block storage cluster.
+        /// </summary>
+        [NameInMap("DbscId")]
+        [Validation(Required=false)]
+        public string DbscId { get; set; }
+
+        /// <summary>
+        /// End timestamp of trend data.
+        /// </summary>
+        [NameInMap("EndTime")]
+        [Validation(Required=false)]
+        public long? EndTime { get; set; }
+
+        /// <summary>
+        /// The time interval （seconds） between data retrieval points.
+        /// </summary>
+        [NameInMap("Period")]
+        [Validation(Required=false)]
+        public int? Period { get; set; }
+
+        /// <summary>
+        /// The region ID of the dedicated block storage cluster.
+        /// </summary>
+        [NameInMap("RegionId")]
+        [Validation(Required=false)]
+        public string RegionId { get; set; }
+
+        /// <summary>
+        /// Start timestamp of trend data.
+        /// </summary>
+        [NameInMap("StartTime")]
+        [Validation(Required=false)]
+        public long? StartTime { get; set; }
+
+    }
+
+}