+++ ebs-20210730/core/Models/RemoveDiskReplicaPairResponse.cs
@@ -10,15 +10,15 @@ namespace AlibabaCloud.SDK.Ebs20210730.Models
 {
     public class RemoveDiskReplicaPairResponse : TeaModel {
         [NameInMap("headers")]
-        [Validation(Required=true)]
+        [Validation(Required=false)]
         public Dictionary<string, string> Headers { get; set; }
 
         [NameInMap("statusCode")]
-        [Validation(Required=true)]
+        [Validation(Required=false)]
         public int? StatusCode { get; set; }
 
         [NameInMap("body")]
-        [Validation(Required=true)]
+        [Validation(Required=false)]
         public RemoveDiskReplicaPairResponseBody Body { get; set; }
 
     }